princeton/src/main/java/princeton/princeton/member/repository/MemberJpaRepository.java
package princeton.princeton.member.repository;

import jakarta.persistence.EntityManager;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Repository;
import princeton.princeton.member.entity.Member;

@Repository
@RequiredArgsConstructor
public class MemberJpaRepository {

    private final EntityManager em;

    public Member save(Member member) {
        em.persist(member);
        return member;
    }

    public Member find(Long memberId) {
        return em.find(Member.class, memberId);
    }
}
